added low quality error/logging facilities esp in regards to environment variables/defaults and added arm support. Explicity toggled on repeaters for scheduled events

+++ gen_callfile.lua
@@ -1,25 +1,63 @@
+function print_warning(msg)
+	--error("ERROR:" .. msg)
+	vim.api.nvim_echo({{"WARNING: " .. msg}}, true , {err = false})
+end
 
-local print_dbg = false --true --false
+-- TODO V rename this into print_dbg_verbose
+local print_dbg_verbose = false -- true -- false --true --false
+
 
 function print_error(msg)
 	--error("ERROR:" .. msg)
-	vim.api.nvim_echo({{"ERROR: "},{msg}}, true , {err = true})
+	vim.api.nvim_echo({{"ERROR: " .. msg}}, true , {err = true})
 end
 
+local print_dbg = false
+local dbg = function(dbg_msg)
+	if print_dbg then
+		print(dbg_msg)
+	end
+end
+-- TODO decide to set the nvim printer only when configured/detected its running in vim
+dbg = function(dbg_msg)
+	if print_dbg then
+		vim.api.nvim_echo({{"\tdbg:" .. dbg_msg}}, true , {err = false})
+	end
+end
+
+function dbg_section_enter(section_name)
+	if print_dbg then
+		vim.api.nvim_echo({{"dbg START section[" .. section_name .."]:" }}, true , {err = false})
+	end
+end
+function dbg_section_leave(section_name)
+	if print_dbg then
+		vim.api.nvim_echo({{"dbg END section[" .. section_name .."]:" }}, true , {err = false})
+	end
+end
+
+
 function run_command(command)
+	dbg_section_enter("run_command")
+	dbg(command)
 	local openPop = assert(io.popen(command))
 	local output = openPop:read('*all')
 	openPop:close()
-	if print_dbg then print(output) end
+	if print_dbg_verbose then print(output) end
+	dbg_section_leave("run_command")
 end
 
 local function write_file(path, content)
+  dbg_section_enter("write_file")
   local f, err = io.open(path, "wb")
   if not f then
+  	  print_error("while writing to file \"" .. err .. "\":" .. err )
 	  return false, err
   end
   f:write(content)
   f:close()
+  dbg("wrote to: " .. path)
+  dbg_section_leave("write_file")
   return true
 end
 
@@ -28,11 +66,14 @@ end
 -- we allow calendar entries to have numbers
 -- parsed out of them
 function normalize_pstn_simple(str)
+  dbg_section_enter("normalize_pstn_simple")
   -- remove all whitespace characters
   str = str:gsub("%s+", "")
   if str:sub(1,1) == "+" then
+    dbg_section_leave("normalize_pstn_simple")
     return "00" .. str:sub(2)
   end
+  dbg_section_leave("normalize_pstn_simple")
   return str
 end
 
@@ -40,7 +81,7 @@ function get_channel(raw_target_str)
   if tostring(raw_target_str):match("^%d%d?%d?%d?$") then
     local channel = "PJSIP/" .. raw_target_str
     -- TODO V turn to vim notify
-    if print_dbg then 
+    if print_dbg_verbose then 
     	print(("Target detected as local extension: %s"):format(raw_target_str))
     end
     return channel, raw_target_str
@@ -48,7 +89,7 @@ function get_channel(raw_target_str)
     local pstn = normalize_pstn_simple(raw_target_str)
     local channel = ("Local/%s@from-internal"):format(pstn)
     -- TODO V turn to vim notify
-    if print_dbg then 
+    if print_dbg_verbose then 
     	print(("Target detected as external number: %s via dialplan"):format(pstn))
     end
     return channel, pstn
@@ -58,6 +99,7 @@ end
 
 
 function pause_to_wav(wav_file_name)
+dbg_section_enter("pause_to_wav")
 run_command(
   table.concat({
     "sox",
@@ -73,16 +115,18 @@ run_command(
     "0.0 0.2"
   }, " ")
 )
+  dbg_section_leave("pause_to_wav")
   return wav_file_name
 
 end
 
-function deploy_files(cfg, wav_file_names, target)
+function deploy_files(cfg, wav_file_names, target,mv)
+	dbg_section_enter("deploy_files")
 	-- this is so you can set e.g 0660 the asterisk group can
 	-- remove them again as well (unless you want to keep them)
 	local group = ""
 	if cfg.group ~= nil then
-		group = "-g asterisk"
+		group = cfg.group
 	end
 
 	local perms = "0640"
@@ -90,19 +134,37 @@ function deploy_files(cfg, wav_file_names, target)
 		perms = cfg.perms
 	end
 	for _,f in pairs(wav_file_names) do
-		run_command(table.concat({
-			"install",
-			"-m",perms,
-			-- "-o", "asterisk",
-			-- TODO set the GROUP!! <disabled as im nonroot rn >
-			-- TODO figure out a way in which asterisks deletes the
-			-- audiofile if the call succeeded
-			group,
-			f,
-			target
-		}, " "))
-		-- TODO install doesn't move like "mv" right ?!
-		os.remove(f)
+		-- had to add this as I observed a semingly random portion
+		-- of calls getting lost ... the underlaying problem seems
+		-- to be a race condition where asterisk spool directory
+		-- is watched by inotify but the "install" command I guess
+		-- first copies the file over AND THEN change the permissions?!
+		-- IF now inotify ends up picking up the file BEFORE the permission
+		-- change syscall is performed by "install" then asterisk complains
+		-- of being unable to read the file and its being deleted.
+		if mv then
+			run_command(string.format("chown :%s %s",group ,f))
+			run_command(string.format("chmod %s %s",perms ,f))
+			run_command(string.format("mv %s %s/", f, target))
+		else 
+			local c = table.concat({
+				"install",
+				"-m",perms,
+				-- "-o", "asterisk",
+				-- TODO figure out a way in which asterisks deletes the
+				-- audiofile if the call succeeded
+				"-g " .. group,
+				f,
+				"--target-directory=" .. target
+			}, " ")
+
+			run_command(c)
+
+			-- TODO install doesn't move like "mv" right ?!
+			dbg("removing \"" .. f .. "\"")
+			os.remove(f)
+		end
+		dbg_section_leave("deploy_files")
 	end
 end
 
@@ -124,18 +186,36 @@ end
 -- sense as a default if it wouldn't work out of the box or notify
 -- the user about missing stuff
 function say_text_to_wav(cfg,raw_text,output_name)
-	local language_model_file = 
-		cfg.language_model_file or "/tmp/lm/en_US-amy-low.onnx"
+	dbg_section_enter("say_text_to_wav")
 
-	-- TODO V maybe allow external overriding
+  	if (cfg.language_model_file == nil) then
+  	        print_error("cfg.language_model_file not provided to say_text_to_wav(), required for callfile generation")
+		-- TODO i don't believe it makes any sense to accept defaults here right?
+		assert(false);
+  	end
+	local language_model_file = cfg.language_model_file
+
 	local language_model_config_file = language_model_file .. ".json"
+	if cfg.language_model_config_file == nil then
+		print_warning("cfg.language_model_config_file unset, defaulting to: " ..
+		language_model_config_file)
+	end
+
 	-- V TODO $(mktemp) as default
-	local temp_audio_files_dir = (cfg.temp_audio_files_dir or "/tmp/tempaudio")
-	.. "/"
+
+	if cfg.temp_audio_files_dir == nil then
+	  print_error("temp_audio_files_dir not provided, required for callfile generation")
+	  assert(true)
+	end
+
+	local temp_audio_files_dir = cfg.temp_audio_files_dir .. "/" -- add "/" for good measure
+	-- TODO i don't think defaults make sense right?!
+	--local temp_audio_files_dir = (cfg.temp_audio_files_dir or "/tmp/tempaudio")
 
 	local temp_speech_file = temp_audio_files_dir .. "/speech.wav"
 	local output_name = temp_audio_files_dir .. output_name .. ".wav"
 	local temp_text_file  =  temp_audio_files_dir .. "temp_text.txt"
+
 	write_file(temp_text_file, raw_text)
 
 	run_command( table.concat({
@@ -168,6 +248,7 @@ function say_text_to_wav(cfg,raw_text,output_name)
 
 	assert(os.remove( temp_speech_file ))
 	assert(os.remove( temp_text_file ))
+	dbg_section_leave("say_text_to_wav")
 	return output_name 
 end
 
@@ -242,6 +323,8 @@ function make_callfile_content(cfg,_target)
     string.format("Context: %s", context),
     string.format("Extension: %s", tostring(extension)),
     string.format("Priority: %s", tostring(priority)),
+    -- V TODO make this optional
+    string.format("Archive: %s", "yes"),
     table.concat(setvar_lines, "\n"),
     ""
   }, "\n")
@@ -249,77 +332,6 @@ end
 
 
 
-
---local dummy_tasks = {
---	{
---		title ="wash the dishes",
---		time = "twelve O two",
---	},
---	{
---		title ="Bring out the trash",
---		time = "fourteen thirtythree",
---	},
---	{
---		title ="make dinner",
---		time = "eighteen thirty",
---	},
---}
-
-
-
--- TODO V TODO use table.getn(tasks) instead ig, this is kinda ghetto
--- 		idk why its not here
-
-function vocalize_tasks(tasks)
-  local vocalized_tasks = {}
-  local _task_count = 0
-  for taskno,task in ipairs(tasks) do
-  	-- TODO V TODO use getn instead ig, this is kinda ghetto
-  	_task_count = _task_count + 1
-  	-- rather use static pauses in the cal file,
-  	-- speeds things up ...
-  	--paused_frag1 = pause_to_wav("pause1.wav")
-  	--paused_frag2 = pause_to_wav("pause2.wav")
-  	task_title_vocal = say_text_to_wav({}, task.title, "task_".. taskno .. "_title")
-  	task_time_vocal = say_text_to_wav({}, task.time, "task_".. taskno .. "_time")
-  	-- task_final_vocal = combine_wavs( 
-  	-- 	{
-  	-- 		task_title_vocal,
-  	-- 		--paused_frag1,
-  	-- 		--paused_frag2,
-  	-- 		task_time_vocal
-  	-- 	})
-  	table.insert( vocalized_tasks, task_title_vocal)
-  	table.insert( vocalized_tasks, task_time_vocal)
-  end
-  return vocalized_tasks
-end
-
---deploy_wavs({ perms = "0660"},vocalized_tasks , dynamic_sounds_dir)
-
---lol_frag = say_text_to_wav({},"L O L", "lol")
---
---
---lole_frag = say_text_to_wav({},"LOLE", "lole")
---
---final_frag = combine_wavs( { lol_frag, paused_frag, lole_frag }, "combined")
-
-
-
-  --local callfile_name = string.format("%s/notify_%s_%d.call", callfile_dir, target, os.time())
-
---print(final_frag)
-
--- dialog_task_ack.wav  dialog_task_try_again.wav  goodbye.wav         greeting_part2.wav
--- dialog_task_ask.wav  gap_200ms.wav              greeting_part1.wav
-
---x = make_callfile_content({
---	task_count = _task_count,
---	dynamic_sounds_dir = dynamic_sounds_dir,
---	static_sounds_dir = "/nix/store/l747zs99y5x0w987z0kxm2rx1pl14341-asterisk_custom_piper_sounds",
---}, "+385<number>")
-
---write_file("callfile_test.cf",x)
 return {
 	say_text_to_wav = say_text_to_wav,
 	make_callfile_content = make_callfile_content,
